Renamed RecurringEventActivity to BookEventActivity
