+++ modality-event/modality-event-frontoffice-activity-booking-plugin/src/main/java/one/modality/event/frontoffice/activities/booking/process/event/BookEventActivity.java
@@ -1,4 +1,4 @@
-package one.modality.event.frontoffice.activities.booking.process.recurring;
+package one.modality.event.frontoffice.activities.booking.process.event;
 
 import dev.webfx.platform.console.Console;
 import dev.webfx.stack.orm.domainmodel.activity.viewdomain.impl.ViewDomainActivityBase;
@@ -9,11 +9,21 @@ import one.modality.event.frontoffice.activities.booking.fx.FXEventAggregate;
 /**
  * @author Bruno Salmon
  */
-public final class RecurringEventActivity extends ViewDomainActivityBase {
+public final class BookEventActivity extends ViewDomainActivityBase {
 
     @Override
     public Node buildUi() {
         return new Text("Recurring Event");
+    }
+
+    @Override
+    protected void updateContextParametersFromRoute() {
+        super.updateContextParametersFromRoute();
+    }
+
+    @Override
+    protected void updateModelFromContextParameters() {
+        super.updateModelFromContextParameters();
     }
 
     @Override
